@@ -1,33 +1,34 @@
 <?php
 
-class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
+class ConceptMappingPropertyValueTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  private $model;
   private $concept;
   private $vocab;
   private $props;
 
   protected function setUp() {
-    require_once 'testconfig.inc';
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('mapping');
     $concepts = $this->vocab->getConceptInfo('http://www.skosmos.skos/mapping/m1', 'en');
     $this->concept = $concepts[0];
     $this->props = $this->concept->getMappingProperties();
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::__construct
    */
   public function testConstructor() {
     $resourcestub = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
-    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $resourcestub, 'skos:exactMatch');
+    $sourcestub = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $resourcestub, $sourcestub, 'skos:exactMatch');
     $this->assertEquals('skos:exactMatch', $mapping->getType());
   }
 
@@ -46,6 +47,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
    * @covers ConceptMappingPropertyValue::queryLabel
    */
   public function testGetLabelResortsToUri() {
+    $mocksource = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $mockres = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $labelmap = array(
       array('en', null),
@@ -58,7 +60,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     );
     $mockres->method('getLiteral')->will($this->returnValueMap($litmap));
     $mockres->method('getUri')->will($this->returnValue('http://thisdoesntexistatalland.sefsf/2j2h4/'));
-    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, null);
+    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, $mocksource, 'skos:exactMatch');
     $this->assertEquals('http://thisdoesntexistatalland.sefsf/2j2h4/', $mapping->getLabel());
   }
 
@@ -67,6 +69,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
    * @covers ConceptMappingPropertyValue::queryLabel
    */
   public function testGetLabelWithAndWithoutLang() {
+    $mocksource = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $mockres = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $labelmap = array(
       array('en', 'english'),
@@ -74,7 +77,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     );
     $mockres->method('label')->will($this->returnValueMap($labelmap));
     $mockres->method('getUri')->will($this->returnValue('http://thisdoesntexistatalland.sefsf/2j2h4/'));
-    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, null);
+    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, $mocksource, 'skos:exactMatch');
     $this->assertEquals('english', $mapping->getLabel('en'));
     $this->assertEquals('default', $mapping->getLabel());
   }
@@ -84,6 +87,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
    * @covers ConceptMappingPropertyValue::queryLabel
    */
   public function testGetLabelWithLiteralAndLang() {
+    $mocksource = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $mockres = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $labelmap = array(
       array('en', null),
@@ -96,7 +100,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     );
     $mockres->method('getLiteral')->will($this->returnValueMap($litmap));
     $mockres->method('getUri')->will($this->returnValue('http://thisdoesntexistatalland.sefsf/2j2h4/'));
-    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, null);
+    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, $mocksource, 'skos:exactMatch');
     $this->assertEquals('english lit', $mapping->getLabel('en'));
     $this->assertEquals('default lit', $mapping->getLabel());
   }
@@ -105,6 +109,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
    * @covers ConceptMappingPropertyValue::getNotation
    */
   public function testGetNotation() {
+    $mocksource = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $mockres = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $mocklit = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
     $mocklit->method('getValue')->will($this->returnValue('666'));
@@ -113,7 +118,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
         array(null,null,null,null),
     );
     $mockres->method('get')->will($this->returnValueMap($map));
-    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, null);
+    $mapping = new ConceptMappingPropertyValue($this->model, $this->vocab, $mockres, $mocksource, 'skos:exactMatch');
     $this->assertEquals(666, $mapping->getNotation());
   }
 
@@ -125,7 +130,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $this->assertInstanceOf('Vocabulary', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getExVocab());
     $this->assertEquals('test', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getExVocab()->getId());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getVocabName
    */
@@ -133,7 +138,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('Test ontology', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getVocabName());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getUri
    */
@@ -141,7 +146,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('http://www.skosmos.skos/test/ta115', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getUri());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getVocab
    */
@@ -149,7 +154,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals($this->vocab, $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getVocab());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getType
    */
@@ -157,13 +162,45 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('skos:exactMatch', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getType());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::__toString
    */
   public function testToString() {
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('Eel', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->__toString());
+  }
+
+  /**
+   * @covers ConceptMappingPropertyValue::asJskos
+   */
+  public function testAsJskos() {
+    $propvals = $this->props['skos:exactMatch']->getValues();
+    $this->assertEquals([
+      'type' => [
+        'skos:exactMatch',
+      ],
+      'toScheme' => [
+        'uri' => 'http://www.skosmos.skos/test/conceptscheme',
+      ],
+      'from' => [
+        'memberSet' => [
+          [
+            'uri' => 'http://www.skosmos.skos/mapping/m1',
+          ]
+        ]
+      ],
+      'to' => [
+        'memberSet' => [
+          [
+            'uri' => 'http://www.skosmos.skos/test/ta115',
+            'prefLabel' => [
+              'en' => 'Eel',
+            ]
+          ]
+        ]
+      ],
+    ], $propvals['Eelhttp://www.skosmos.skos/test/ta115']->asJskos());
   }
 
 }
