@@ -1,14 +1,20 @@
 <?php
 
-class VocabularyConfigTest extends PHPUnit_Framework_TestCase
+class VocabularyConfigTest extends PHPUnit\Framework\TestCase
 {
-  
-  private $model; 
+  /** @var Model */
+  private $model;
 
+  /**
+   * @covers VocabularyConfig::getConfig
+   * @throws Exception
+   */
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
+    $this->assertNotNull($this->model->getVocabulary('test')->getConfig()->getPlugins(), "The PluginRegister of the model was not initialized!");
   }
 
   /**
@@ -19,7 +25,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $actual = $vocab->getConfig()->getIndexClasses();
     $this->assertEquals(array(), $actual);
   }
-  
+
   /**
    * @covers VocabularyConfig::getIndexClasses
    */
@@ -29,7 +35,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $expected = array('http://www.skosmos.skos/test-meta/TestClass','http://www.skosmos.skos/test-meta/TestClass2');
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers VocabularyConfig::getLanguages
    */
@@ -38,7 +44,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $langs = $vocab->getConfig()->getLanguages();
     $this->assertEquals(2, sizeof($langs));
   }
-  
+
   /**
    * @covers VocabularyConfig::getFeedbackRecipient
    */
@@ -47,7 +53,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $email = $vocab->getConfig()->getFeedbackRecipient();
     $this->assertEquals('developer@vocabulary.org', $email);
   }
-  
+
   /**
    * @covers VocabularyConfig::getExternalResourcesLoading
    */
@@ -56,7 +62,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $external = $vocab->getConfig()->getExternalResourcesLoading();
     $this->assertTrue($external);
   }
-  
+
   /**
    * @covers VocabularyConfig::getDefaultSidebarView
    */
@@ -74,7 +80,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $default = $vocab->getConfig()->getDefaultSidebarView();
     $this->assertEquals('alphabetical', $default);
   }
-  
+
   /**
    * @covers VocabularyConfig::getShowLangCodes
    */
@@ -83,7 +89,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $codes = $vocab->getConfig()->getShowLangCodes();
     $this->assertTrue($codes);
   }
-  
+
   /**
    * @covers VocabularyConfig::getShowLangCodes
    */
@@ -92,7 +98,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $codes = $vocab->getConfig()->getShowLangCodes();
     $this->assertFalse($codes);
   }
-  
+
   /**
    * @covers VocabularyConfig::getDefaultLanguage
    * @covers VocabularyConfig::getLiteral
@@ -102,16 +108,16 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $lang = $vocab->getConfig()->getDefaultLanguage();
     $this->assertEquals('en', $lang);
   }
-  
+
   /**
    * @covers VocabularyConfig::getDefaultLanguage
-   * @expectedException PHPUnit_Framework_Error 
+   * @expectedException PHPUnit\Framework\Error\Error
    */
   public function testGetDefaultLanguageWhenNotSet() {
     $vocab = $this->model->getVocabulary('testdiff');
     $lang = $vocab->getConfig()->getDefaultLanguage();
   }
-  
+
   /**
    * @covers VocabularyConfig::getAlphabeticalFull
    */
@@ -120,7 +126,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $boolean = $vocab->getConfig()->getAlphabeticalFull();
     $this->assertEquals(true, $boolean);
   }
-  
+
   /**
    * @covers VocabularyConfig::getAlphabeticalFull
    */
@@ -129,7 +135,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $boolean = $vocab->getConfig()->getAlphabeticalFull();
     $this->assertEquals(false, $boolean);
   }
-  
+
   /**
    * @covers VocabularyConfig::getShortName
    * @covers VocabularyConfig::getLiteral
@@ -139,7 +145,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $name = $vocab->getConfig()->getShortName();
     $this->assertEquals('Test short', $name);
   }
-  
+
   /**
    * @covers VocabularyConfig::getShortName
    */
@@ -148,29 +154,29 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $name = $vocab->getConfig()->getShortName();
     $this->assertEquals('testdiff', $name);
   }
-  
+
   /**
    * @covers VocabularyConfig::getDataURLs
    */
   public function testGetDataURLs() {
     $vocab = $this->model->getVocabulary('groups');
     $url = $vocab->getConfig()->getDataURLs();
-    ksort($url); // sort by mime type to make order deterministic 
+    ksort($url); // sort by mime type to make order deterministic
     $this->assertEquals(array(
-        'application/rdf+xml' => 'http://skosmos.skos/dump/test/groups', 
-        'text/turtle' => 'http://skosmos.skos/dump/test/groups.ttl', 
+        'application/rdf+xml' => 'http://skosmos.skos/dump/test/groups',
+        'text/turtle' => 'http://skosmos.skos/dump/test/groups.ttl',
       ), $url);
   }
 
   /**
    * @covers VocabularyConfig::getDataURLs
-   * @expectedException PHPUnit_Framework_Error_Warning
+   * @expectedException PHPUnit\Framework\Error\Warning
    */
   public function testGetDataURLsNotGuessable() {
     $vocab = $this->model->getVocabulary('test');
     $url = $vocab->getConfig()->getDataURLs();
   }
-  
+
   /**
    * @covers VocabularyConfig::getDataURLs
    */
@@ -188,7 +194,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $uri = $vocab->getConfig()->getGroupClassURI();
     $this->assertEquals('http://www.w3.org/2004/02/skos/core#Collection', $uri);
   }
-  
+
   /**
    * @covers VocabularyConfig::getGroupClassURI
    */
@@ -197,7 +203,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $uri = $vocab->getConfig()->getGroupClassURI();
     $this->assertEquals(null, $uri);
   }
-  
+
   /**
    * @covers VocabularyConfig::getArrayClassURI
    */
@@ -206,7 +212,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $uri = $vocab->getConfig()->getArrayClassURI();
     $this->assertEquals('http://purl.org/iso25964/skos-thes#ThesaurusArray', $uri);
   }
-  
+
   /**
    * @covers VocabularyConfig::getArrayClassURI
    */
@@ -215,7 +221,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $uri = $vocab->getConfig()->getArrayClassURI();
     $this->assertEquals(null, $uri);
   }
-  
+
   /**
    * @covers VocabularyConfig::getShowHierarchy
    */
@@ -225,7 +231,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(true, $uri);
   }
 
- 
+
   /**
    * @covers VocabularyConfig::getShowHierarchy
    */
@@ -234,7 +240,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $uri = $vocab->getConfig()->getShowHierarchy();
     $this->assertEquals(false, $uri);
   }
-  
+
   /**
    * @covers VocabularyConfig::getShowHierarchy
    */
@@ -243,7 +249,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $uri = $vocab->getConfig()->getShowHierarchy();
     $this->assertEquals(false, $uri);
   }
-  
+
   /**
    * @covers VocabularyConfig::getAdditionalSearchProperties
    */
@@ -260,7 +266,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $vocab = $this->model->getVocabulary('dates');
     $this->assertEquals(true, $vocab->getConfig()->hasMultiLingualProperty('skos:altLabel'));
   }
-  
+
   /**
    * @covers VocabularyConfig::hasMultiLingualProperty
    */
@@ -268,7 +274,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $vocab = $this->model->getVocabulary('dates');
     $this->assertEquals(false, $vocab->getConfig()->hasMultiLingualProperty('skos:exactMatch'));
   }
-  
+
   /**
    * @covers VocabularyConfig::getTitle
    */
@@ -346,7 +352,7 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
     $vocab = $this->model->getVocabulary('test');
     $this->assertEquals(array(0 => array('uri' => 'http://publications.europa.eu/resource/authority/dataset-type/ONTOLOGY', 'prefLabel' => 'Ontology')), $vocab->getConfig()->getTypes('en'));
   }
-  
+
   /**
    * @covers VocabularyConfig::getShowDeprecated
    */
@@ -361,5 +367,90 @@ class VocabularyConfigTest extends PHPUnit_Framework_TestCase
   public function testGetTypesWhenNotSet() {
     $vocab = $this->model->getVocabulary('testdiff');
     $this->assertEquals(array(), $vocab->getConfig()->getTypes('en'));
+  }
+
+  /**
+   * @covers VocabularyConfig::getLanguageOrder
+   */
+  public function testGetLanguageOrderNotSet() {
+    $vocab = $this->model->getVocabulary('test');
+    $this->assertEquals(array('en'), $vocab->getConfig()->getLanguageOrder('en'));
+  }
+
+  /**
+   * @covers VocabularyConfig::getLanguageOrder
+   */
+  public function testGetLanguageOrder() {
+    $vocab = $this->model->getVocabulary('subtag');
+    $this->assertEquals(array('en', 'fr', 'de', 'sv'), $vocab->getConfig()->getLanguageOrder('en'));
+    $this->assertEquals(array('fi', 'fr', 'de', 'sv', 'en'), $vocab->getConfig()->getLanguageOrder('fi'));
+  }
+
+  /**
+   * @covers VocabularyConfig::showAlphabeticalIndex
+   */
+  public function testShowAlphabeticalIndex() {
+    $vocab = $this->model->getVocabulary('testdiff');
+    $this->assertTrue($vocab->getConfig()->showAlphabeticalIndex());
+  }
+
+  /**
+   * @covers VocabularyConfig::showNotation
+   */
+  public function testShowNotation() {
+    $vocab = $this->model->getVocabulary('test');
+    $this->assertTrue($vocab->getConfig()->showNotation());
+  }
+
+  /**
+   * @covers VocabularyConfig::getId
+   */
+  public function testGetId() {
+    $vocab = $this->model->getVocabulary('testdiff');
+    $this->assertEquals('testdiff' , $vocab->getConfig()->getId());
+  }
+
+  /**
+   * @covers VocabularyConfig::getMainConceptSchemeURI
+   */
+  public function testGetMainConceptSchemeURI() {
+    $vocab = $this->model->getVocabulary('testdiff');
+    $this->assertEquals('http://www.skosmos.skos/testdiff#conceptscheme' , $vocab->getConfig()->getMainConceptSchemeURI());
+    $vocab = $this->model->getVocabulary('test');
+    $this->assertNull($vocab->getConfig()->getMainConceptSchemeURI());
+  }
+
+  /**
+   * @covers VocabularyConfig::getExtProperties
+   */
+  public function testGetExtProperties() {
+    $vocab = $this->model->getVocabulary('cbd');
+    $this->assertEquals(4 , count($vocab->getConfig()->getExtProperties()));
+  }
+
+  /**
+   * @covers VocabularyConfig::getMarcSourceCode
+   */
+  public function testGetMarcSourceCode() {
+    $vocab = $this->model->getVocabulary('test');
+    $this->assertEquals("ysa/fi" , $vocab->getConfig()->getMarcSourceCode("fi"));
+  }
+
+ /**
+   * @covers VocabularyConfig::getMarcSourceCode
+   */
+  public function testGetMarcSourceCodeWithoutLang() {
+    $vocab = $this->model->getVocabulary('multiple-schemes');
+    $this->assertEquals("ysa/gen" , $vocab->getConfig()->getMarcSourceCode("fi"));
+  }
+
+  /**
+   * @covers VocabularyConfig::getUseModifiedDate
+   */
+  public function testGetVocabularyUseModifiedDate() {
+    $vocab = $this->model->getVocabulary('http304');
+    $this->assertEquals(true , $vocab->getConfig()->getUseModifiedDate());
+    $vocab = $this->model->getVocabulary('http304disabled');
+    $this->assertEquals(false , $vocab->getConfig()->getUseModifiedDate());
   }
 }
