@@ -1,16 +1,17 @@
 <?php
 
-require_once('model/Model.php');
-
-class VocabularyTest extends PHPUnit_Framework_TestCase
+class VocabularyTest extends \PHPUnit\Framework\TestCase
 {
-  
-  private $model; 
+  /**
+   * @var Model
+   */
+  private $model;
 
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
   }
 
   /**
@@ -21,7 +22,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $id = $vocab->getId();
     $this->assertEquals('test', $id);
   }
-  
+
   /**
    * @covers Vocabulary::getTitle
    */
@@ -30,14 +31,14 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $title = $vocab->getTitle();
     $this->assertEquals('Test ontology', $title);
   }
-  
+
   /**
    * @covers Vocabulary::getEndpoint
    */
   public function testGetEndpoint() {
     $vocab = $this->model->getVocabulary('testdiff');
     $endpoint = $vocab->getEndpoint();
-    $this->assertEquals('http://localhost:3030/ds/sparql', $endpoint);
+    $this->assertEquals('http://localhost:13030/ds/sparql', $endpoint);
   }
 
   /**
@@ -48,7 +49,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $graph = $vocab->getGraph();
     $this->assertEquals('http://www.skosmos.skos/testdiff/', $graph);
   }
-  
+
   /**
    * @covers Vocabulary::getSparql
    */
@@ -57,7 +58,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $sparql = $vocab->getSparql();
     $this->assertInstanceOf('GenericSparql', $sparql);
   }
-  
+
   /**
    * @covers Vocabulary::getSparql
    */
@@ -66,7 +67,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $sparql = $vocab->getSparql();
     $this->assertInstanceOf('JenaTextSparql', $sparql);
   }
-  
+
   /**
    * @covers Vocabulary::getUriSpace
    */
@@ -75,7 +76,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $sparql = $vocab->getUriSpace();
     $this->assertEquals('http://www.skosmos.skos/onto/testdiff#', $sparql);
   }
-  
+
   /**
    * @covers Vocabulary::getLocalName
    */
@@ -84,7 +85,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $name = $vocab->getLocalName('http://www.skosmos.skos/onto/testdiff#concept23');
     $this->assertEquals('concept23', $name);
   }
-  
+
   /**
    * @covers Vocabulary::getConceptURI
    */
@@ -93,7 +94,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $name = $vocab->getConceptURI('concept23');
     $this->assertEquals('http://www.skosmos.skos/onto/testdiff#concept23', $name);
   }
-  
+
   /**
    * @covers Vocabulary::getConceptURI
    */
@@ -102,7 +103,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $name = $vocab->getConceptURI('http://www.skosmos.skos/onto/testdiff#concept23');
     $this->assertEquals('http://www.skosmos.skos/onto/testdiff#concept23', $name);
   }
-  
+
   /**
    * @covers Vocabulary::getDefaultConceptScheme
    */
@@ -111,7 +112,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $cs = $vocab->getDefaultConceptScheme();
     $this->assertEquals('http://www.skosmos.skos/testdiff#conceptscheme', $cs);
   }
-  
+
   /**
    * @covers Vocabulary::getDefaultConceptScheme
    */
@@ -120,7 +121,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $cs = $vocab->getDefaultConceptScheme();
     $this->assertEquals('http://www.skosmos.skos/test/conceptscheme', $cs);
   }
-  
+
   /**
    * @covers Vocabulary::getConceptSchemes
    */
@@ -132,7 +133,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
       $this->assertEquals('Test conceptscheme', $label['label']);
     }
   }
-  
+
   /**
    * @covers Vocabulary::getLabelStatistics
    */
@@ -143,7 +144,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
       $this->assertEquals(11, $labels['skos:prefLabel']);
     }
   }
-  
+
   /**
    * @covers Vocabulary::getStatistics
    */
@@ -152,7 +153,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $stats = $vocab->getStatistics();
     $this->assertEquals(16, $stats['http://www.w3.org/2004/02/skos/core#Concept']['count']);
   }
-  
+
   /**
    * @covers Vocabulary::listConceptGroups
    */
@@ -162,7 +163,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $expected = array (0 => array ('uri' => 'http://www.skosmos.skos/groups/fish', 'hasMembers' => true, 'childGroups' => array('http://www.skosmos.skos/groups/sub'), 'prefLabel' => 'Fish'), 1 => array ('uri' => 'http://www.skosmos.skos/groups/fresh', 'hasMembers' => true, 'prefLabel' => 'Freshwater fish'), 2 => array ('uri' => 'http://www.skosmos.skos/groups/salt', 'hasMembers' => true, 'prefLabel' => 'Saltwater fish'),3 => array ('uri' => 'http://www.skosmos.skos/groups/sub', 'hasMembers' => true, 'prefLabel' => 'Submarine-like fish'));
     $this->assertEquals($expected, $cgroups);
   }
-  
+
   /**
    * @covers Vocabulary::listConceptGroupContents
    */
@@ -172,7 +173,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $expected = array (0 => array ('uri' => 'http://www.skosmos.skos/groups/ta113','isSuper' => false,'hasMembers' => false,'type' => array (0 => 'skos:Concept'),'prefLabel' => 'Flatfish'));
     $this->assertEquals($expected, $cgroups);
   }
-  
+
   /**
    * @covers Vocabulary::getAlphabet
    */
@@ -220,7 +221,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $info = $vocab->getInfo();
     $this->assertEquals(array('2016-02-11j14:05:26Z someone@skosmos.skos (r1197)'), $info['owl:versionInfo']);
   }
-  
+
   /**
    * @covers Vocabulary::getInfo
    */
@@ -240,7 +241,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('Grouped fish', $concepts[0]['prefLabel']);
     $this->assertEquals('Guppy', $concepts[1]['prefLabel']);
   }
-  
+
   /**
   * @covers Vocabulary::searchConceptsAlphabetical
   */
@@ -249,7 +250,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
       $concepts = $vocab->searchConceptsAlphabetical('T', null, null, 'en');
       $this->assertCount(0, $concepts);
   }
-  
+
   /**
    * @covers Vocabulary::searchConceptsAlphabetical
    */
@@ -295,24 +296,24 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
    * @covers Vocabulary::getCrumbs
    */
   public function testGetBreadCrumbs() {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $resource = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
-    $vocabstub = $this->getMock('Vocabulary', array('getConceptTransitiveBroaders'), array($model, $resource));
+    $vocabstub = $this->getMockBuilder('Vocabulary')->setMethods(array('getConceptTransitiveBroaders'))->setConstructorArgs(array($model, $resource))->getMock();
     $vocabstub->method('getConceptTransitiveBroaders')->willReturn(array ( 'http://www.yso.fi/onto/yso/p4762' => array ( 'label' => 'objects', ), 'http://www.yso.fi/onto/yso/p1674' => array ( 'label' => 'physical whole', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p4762', ), ), 'http://www.yso.fi/onto/yso/p14606' => array ( 'label' => 'layers', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p1674', ), ), ));
     $result = $vocabstub->getBreadCrumbs('en', 'http://www.yso.fi/onto/yso/p14606');
-    foreach($result['breadcrumbs'][0] as $crumb)    
+    foreach($result['breadcrumbs'][0] as $crumb)
       $this->assertInstanceOf('Breadcrumb', $crumb);
   }
-  
+
   /**
    * @covers Vocabulary::getBreadCrumbs
    * @covers Vocabulary::combineCrumbs
    * @covers Vocabulary::getCrumbs
    */
   public function testGetBreadCrumbsShortening() {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $resource = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
-    $vocabstub = $this->getMock('Vocabulary', array('getConceptTransitiveBroaders'), array($model, $resource));
+    $vocabstub = $this->getMockBuilder('Vocabulary')->setMethods(array('getConceptTransitiveBroaders'))->setConstructorArgs(array($model, $resource))->getMock();
     $vocabstub->method('getConceptTransitiveBroaders')->willReturn(array ( 'http://www.yso.fi/onto/yso/p4762' => array ( 'label' => 'objects', ), 'http://www.yso.fi/onto/yso/p13871' => array ( 'label' => 'thai language', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p10834', ), ), 'http://www.yso.fi/onto/yso/p556' => array ( 'label' => 'languages', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p2881', ), ), 'http://www.yso.fi/onto/yso/p8965' => array ( 'label' => 'Sino-Tibetan languages', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p556', ), ), 'http://www.yso.fi/onto/yso/p3358' => array ( 'label' => 'systems', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p4762', ), ), 'http://www.yso.fi/onto/yso/p10834' => array ( 'label' => 'Tai languages', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p8965', ), ), 'http://www.yso.fi/onto/yso/p2881' => array ( 'label' => 'cultural systems', 'direct' => array ( 0 => 'http://www.yso.fi/onto/yso/p3358', ), ), ) );
     $result = $vocabstub->getBreadCrumbs('en', 'http://www.yso.fi/onto/yso/p13871');
     $this->assertEquals(6, sizeof($result['breadcrumbs'][0]));
@@ -324,10 +325,10 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
    * @covers Vocabulary::getCrumbs
    */
   public function testGetBreadCrumbsCycle() {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $vocab = $model->getVocabulary('cycle');
     $result = $vocab->getBreadCrumbs('en', 'http://www.skosmos.skos/cycle/ta4');
-    foreach ($result['breadcrumbs'][0] as $crumb)    
+    foreach ($result['breadcrumbs'][0] as $crumb)
       $this->assertInstanceOf('Breadcrumb', $crumb);
   }
 
@@ -339,7 +340,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $prop = $vocab->getTopConcepts();
     $this->assertEquals(array (0 => array ('uri' => 'http://www.skosmos.skos/test/ta1','label' => 'Fish','hasChildren' => true, 'topConceptOf' => 'http://www.skosmos.skos/test/conceptscheme')), $prop);
   }
-  
+
   /**
    * @covers Vocabulary::getChangeList
    */
@@ -350,7 +351,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(array('December 2011', 'February 2010', 'January 2000'), array_keys($months));
     $this->assertEquals($expected, $months['February 2010']);
   }
-  
+
   /**
    * @covers Vocabulary::verifyVocabularyLanguage
    */
@@ -359,7 +360,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('en', $vocab->verifyVocabularyLanguage('en'));
     $this->assertEquals('en', $vocab->verifyVocabularyLanguage('de'));
   }
-  
+
   /**
    * @covers Vocabulary::getShortName
    */
@@ -367,7 +368,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $vocab = $this->model->getVocabulary('test');
     $this->assertEquals('Test short', $vocab->getShortName());
   }
-  
+
   /**
    * @covers Vocabulary::getConceptLabel
    */
@@ -376,7 +377,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $lits = $vocab->getConceptLabel('http://www.skosmos.skos/test/ta112', 'en');
     $this->assertEquals('Carp', $lits['en']->getValue());
   }
-  
+
   /**
    * @covers Vocabulary::getConfig
    */
@@ -384,7 +385,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $vocab = $this->model->getVocabulary('test');
     $this->assertInstanceOf('VocabularyConfig', $vocab->getConfig());
   }
-  
+
   /**
    * @covers Vocabulary::listConceptGroups
    */
@@ -394,7 +395,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $vocab = new Vocabulary($this->model, $mockres);
     $this->assertEquals(array(), $vocab->listConceptGroups());
   }
-  
+
   /**
    * @covers Vocabulary::listConceptGroupContents
    */
@@ -404,7 +405,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $vocab = new Vocabulary($this->model, $mockres);
     $this->assertEquals(array(), $vocab->listConceptGroups());
   }
-  
+
   /**
    * @covers Vocabulary::getConceptNarrowers
    */
@@ -414,7 +415,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(array('http://www.skosmos.skos/test/ta121' => array('label' => 'Crucian carp')), $narrowers);
     $this->assertEquals(1, sizeof($narrowers));
   }
-  
+
   /**
    * @covers Vocabulary::getConceptBroaders
    */
@@ -424,7 +425,7 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(array('http://www.skosmos.skos/test/ta1' => array('label' => 'Fish')), $broaders);
     $this->assertEquals(1, sizeof($broaders));
   }
-  
+
   /**
    * @covers Vocabulary::getConceptInfo
    */
@@ -435,5 +436,18 @@ class VocabularyTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, sizeof($concept));
   }
 
+  /**
+   * @covers Vocabulary::getConceptScheme
+   */
+  public function testGetConceptScheme() {
+    $vocab = $this->model->getVocabulary('http304');
+    $conceptSchemeUri = $vocab->getDefaultConceptScheme();
+    $conceptScheme = $vocab->getConceptScheme($conceptSchemeUri);
+    print($conceptSchemeUri);
+    $this->assertEquals(
+      "Test Main Concept Scheme",
+      $conceptScheme->getLiteral($conceptSchemeUri,"skos:prefLabel")
+    );
+  }
 }
 
