@@ -7,7 +7,7 @@ class Controller
 {
     /**
      * The controller has to know the model to access the data stored there.
-     * @param $model contains the Model object.
+     * @var Model $model contains the Model object.
      */
     public $model;
 
@@ -21,7 +21,7 @@ class Controller
     public function __construct($model)
     {
         $this->model = $model;
-        $this->negotiator = new \Negotiation\FormatNegotiator();
+        $this->negotiator = new \Negotiation\Negotiator();
 
         // Specify the location of the translation tables
         bindtextdomain('skosmos', 'resource/translations');
@@ -48,6 +48,7 @@ class Controller
     {
         if (array_key_exists($lang, $this->languages)) {
             $locale = $this->languages[$lang]['locale'];
+            putenv("LANGUAGE=$locale");
             putenv("LC_ALL=$locale");
             setlocale(LC_ALL, $locale);
         } else {
@@ -73,7 +74,7 @@ class Controller
             }
             return $format;
         }
-        
+
         // if there was no proposed format, negotiate a suitable format
         header('Vary: Accept'); // inform caches that a decision was made based on Accept header
         $best = $this->negotiator->getBest($accept, $choices);
@@ -94,7 +95,7 @@ class Controller
         $base_url = str_replace('/controller', '/', $base_url);
         $protocol = filter_input(INPUT_SERVER, 'HTTPS', FILTER_SANITIZE_STRING) === null ? 'http' : 'https';
         $port = filter_input(INPUT_SERVER, 'SERVER_PORT', FILTER_SANITIZE_STRING);
-        $disp_port = ($protocol == 'http' && $port == 80 || $protocol == 'https' && $port == 443) ? '' : ":$port";
+        $disp_port = ($port == 80 || $port == 443) ? '' : ":$port";
         $domain = filter_input(INPUT_SERVER, 'SERVER_NAME', FILTER_SANITIZE_STRING);
         $full_url = "$protocol://{$domain}{$disp_port}{$base_url}";
         return $full_url;
@@ -116,5 +117,56 @@ class Controller
         header("HTTP/1.0 $code $status");
         header("Content-type: text/plain; charset=utf-8");
         echo "$code $status : $message";
+    }
+
+    /**
+     * If the $modifiedDate is a valid DateTime, and if the $_SERVER variable contains the right info, and
+     * if the $modifiedDate is not more recent than the latest value in $_SERVER, then this function sets the
+     * HTTP 304 not modified and returns true..
+     *
+     * If the $modifiedDate is still valid, then it sets the Last-Modified header, to be used by the browser for
+     * subsequent requests, and returns false.
+     *
+     * Otherwise, it returns false.
+     *
+     * @param DateTime $modifiedDate the last modified date to be compared against server's modified since information
+     * @return bool whether it sent the HTTP 304 not modified headers or not (useful for sending the response without
+     *              further actions)
+     */
+    protected function sendNotModifiedHeader($modifiedDate): bool
+    {
+        if ($modifiedDate) {
+            $ifModifiedSince = $this->getIfModifiedSince();
+            $this->sendHeader("Last-Modified: " . $modifiedDate->format('Y-m-d H:i:s'));
+            if (!is_null($ifModifiedSince) && $ifModifiedSince >= $modifiedDate) {
+                $this->sendHeader("HTTP/1.0 304 Not Modified");
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * @return DateTime|null a DateTime object if the value exists in the $_SERVER variable, null otherwise
+     */
+    protected function getIfModifiedSince()
+    {
+        $ifModifiedSince = null;
+        if (isset($_SERVER["HTTP_IF_MODIFIED_SINCE"])) {
+            // example value set by a browser: "2019-04-13 08:28:23"
+            $ifModifiedSince = DateTime::createFromFormat("Y-m-d H:i:s", $_SERVER["HTTP_IF_MODIFIED_SINCE"]);
+        }
+        return $ifModifiedSince;
+    }
+
+    /**
+     * Sends HTTP headers. Simply calls PHP built-in header function. But being
+     * a function here, it can easily be tested/mocked.
+     *
+     * @param $header string header to be sent
+     */
+    protected function sendHeader($header)
+    {
+        header($header);
     }
 }
