@@ -19,19 +19,25 @@ class EntityController extends Controller
         $url = $baseurl . "rest/v1/$vocid/data?$query";
         $this->redirect303($url);
     }
-    
+
     private function redirectWeb($vocab, $uri)
     {
         $baseurl = $this->getBaseHref();
         $vocid = $vocab->getId();
         $localname = $vocab->getLocalName($uri);
-        if ($localname !== $uri && $localname === urlencode($localname)) {
-            // the URI can be shortened
-            $url = $baseurl . "$vocid/page/$localname";
+
+        if (!$localname) {
+            $url = $baseurl . "$vocid/";
         } else {
-            // must use full URI
-            $query = http_build_query(array('uri'=>$uri));
-            $url = $baseurl . "$vocid/page/?" . $query;
+
+   	        if ($localname !== $uri && $localname === urlencode($localname)) {
+   	        // the URI can be shortened
+            $url = $baseurl . "$vocid/page/$localname";
+            } else {
+                // must use full URI
+   	            $query = http_build_query(array('uri'=>$uri));
+                $url = $baseurl . "$vocid/page/?" . $query;
+	        }
         }
         $this->redirect303($url);
     }
@@ -65,7 +71,7 @@ class EntityController extends Controller
         $requestedFormat = $request->getQueryParam('format');
 
         $targetFormat = $this->negotiateFormat($supportedFormats, $request->getServerConstant('HTTP_ACCEPT'), $requestedFormat);
-        
+
         if (in_array($targetFormat, $restFormats)) {
             $this->redirectREST($request->getVocab(), $request->getUri(), $targetFormat);
         } else {
