@@ -1,5 +1,7 @@
 <?php
 
+use EasyRdf\Resource;
+
 /**
  * Class for handling concept property values.
  */
@@ -7,12 +9,24 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
 {
     /** property type */
     private $type;
+    private $source;
     private $clang;
     private $labelcache;
 
-    public function __construct($model, $vocab, $resource, $prop, $clang = '')
+    /**
+     * ConceptMappingPropertyValue constructor.
+     *
+     * @param Model $model
+     * @param Vocabulary $vocab  Target vocabulary
+     * @param Resource $target   Target concept resource
+     * @param Resource $source   Source concept resource
+     * @param string $prop       Mapping property
+     * @param ?string $clang     Preferred label language (nullable)
+     */
+    public function __construct(Model $model, Vocabulary $vocab, Resource $target, Resource $source, string $prop, $clang = '')
     {
-        parent::__construct($model, $vocab, $resource);
+        parent::__construct($model, $vocab, $target);
+        $this->source = $source;
         $this->type = $prop;
         $this->clang = $clang;
         $this->labelcache = array();
@@ -30,7 +44,7 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
         return $this->type;
     }
 
-    public function getLabel($lang = '')
+    public function getLabel($lang = '', $queryExVocabs = true)
     {
         if (isset($this->labelcache[$lang])) {
             return $this->labelcache[$lang];
@@ -41,25 +55,22 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
         return $label;
     }
 
-    private function queryLabel($lang)
+    private function queryLabel($lang = '', $queryExVocabs = true)
     {
         if ($this->clang) {
             $lang = $this->clang;
         }
 
-        $exvocab = $this->model->guessVocabularyFromURI($this->resource->getUri());
 
-        if ($this->resource->label($lang) !== null) { // current language
-            return $this->resource->label($lang);
-        } elseif ($this->resource->label() !== null) { // any language
-            return $this->resource->label();
-        } elseif ($this->resource->getLiteral('rdf:value', $lang) !== null) { // current language
-            return $this->resource->getLiteral('rdf:value', $lang);
-        } elseif ($this->resource->getLiteral('rdf:value') !== null) { // any language
-            return $this->resource->getLiteral('rdf:value');
+        $label = $this->getResourceLabel($this->resource, $lang);
+        if ($label) {
+            return $label;
         }
 
-        // if the resource is from a another vocabulary known by the skosmos instance
+        // if multiple vocabularies are found, the following method will return in priority the current vocabulary of the mapping
+        $exvocab = $queryExVocabs ? $this->model->guessVocabularyFromURI($this->resource->getUri(), $this->vocab->getId()) : null;
+
+        // if the resource is from another vocabulary known by the skosmos instance
         if ($exvocab) {
             $label = $this->getExternalLabel($exvocab, $this->getUri(), $lang) ? $this->getExternalLabel($exvocab, $this->getUri(), $lang) : $this->getExternalLabel($exvocab, $this->getUri(), $exvocab->getConfig()->getDefaultLanguage());
             if ($label) {
@@ -72,6 +83,24 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
         return $label;
     }
 
+    private function getResourceLabel($res, $lang = '') {
+
+        if ($this->clang) {
+            $lang = $this->clang;
+        }
+
+        if ($res->label($lang) !== null) { // current language
+            return $res->label($lang);
+        } elseif ($res->label() !== null) { // any language
+            return $res->label();
+        } elseif ($res->getLiteral('rdf:value', $lang) !== null) { // current language
+            return $res->getLiteral('rdf:value', $lang);
+        } elseif ($res->getLiteral('rdf:value') !== null) { // any language
+            return $res->getLiteral('rdf:value');
+        }
+        return null;
+    }
+
     public function getUri()
     {
         return $this->resource->getUri();
@@ -79,7 +108,7 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
 
     public function getExVocab()
     {
-        $exvocab = $this->model->guessVocabularyFromURI($this->getUri());
+        $exvocab = $this->model->guessVocabularyFromURI($this->getUri(), $this->vocab->getId());
         return $exvocab;
     }
 
@@ -88,23 +117,39 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
         return $this->vocab;
     }
 
-    public function getVocabName()
+    public function getVocabName($lang = '')
     {
-        $exvocab = $this->model->guessVocabularyFromURI($this->resource->getUri());
-        if ($exvocab) {
-            return $exvocab->getTitle();
+
+        if ($this->clang) {
+            $lang = $this->clang;
         }
+
+        // if multiple vocabularies are found, the following method will return in priority the current vocabulary of the mapping
+        $exvocab = $this->model->guessVocabularyFromURI($this->resource->getUri(), $this->vocab->getId());
+        if ($exvocab) {
+            return $exvocab->getTitle($lang);
+        }
+
         // @codeCoverageIgnoreStart
         $scheme = $this->resource->get('skos:inScheme');
         if ($scheme) {
             $schemeResource = $this->model->getResourceFromUri($scheme->getUri());
-            if ($schemeResource && $schemeResource->label()) {
-                return $schemeResource->label()->getValue();
+            if ($schemeResource) {
+                $schemaName = $this->getResourceLabel($schemeResource);
+                if ($schemaName) {
+                    //var_dump($schemaName);
+                    return $schemaName;
+                }
             }
         }
         // got a label for the concept, but not the scheme - use the host name as scheme label
         return parse_url($this->resource->getUri(), PHP_URL_HOST);
         // @codeCoverageIgnoreEnd
+    }
+
+    public function isExternal() {
+        $propertyUris = $this->resource->propertyUris();
+        return empty($propertyUris);
     }
 
     public function getNotation()
@@ -122,4 +167,67 @@ class ConceptMappingPropertyValue extends VocabularyDataObject
         return null;
     }
 
+    /**
+     * Return the mapping as a JSKOS-compatible array.
+     * @return array
+     */
+    public function asJskos($queryExVocabs = true)
+    {
+        $ret = [
+            'type' => [$this->type],
+            'from' => [
+                'memberSet' => [
+                    [
+                        'uri' => (string) $this->source->getUri(),
+                    ]
+                ]
+            ],
+            'to' => [
+                'memberSet' => [
+                    [
+                        'uri' => (string) $this->getUri()
+                    ]
+                ]
+            ]
+        ];
+
+        $fromScheme = $this->vocab->getDefaultConceptScheme();
+        if (isset($fromScheme)) {
+            $ret['fromScheme'] = [
+                'uri' => (string) $fromScheme,
+            ];
+        }
+
+        $exvocab = $this->getExvocab();
+        if (isset($exvocab)) {
+            $ret['toScheme'] = [
+                'uri' => (string) $exvocab->getDefaultConceptScheme(),
+            ];
+        }
+
+        $notation = $this->getNotation();
+        if (isset($notation)) {
+            $ret['to']['memberSet'][0]['notation'] = (string) $notation;
+        }
+
+        $label = $this->getLabel(null, $queryExVocabs);
+        if (isset($label)) {
+            if (is_string($label)) {
+                list($labelLang, $labelValue) = ['-', $label];
+            } else {
+                list($labelLang, $labelValue) = [$label->getLang(), $label->getValue()];
+            }
+            if ($labelValue != $this->getUri()) {
+                // The `queryLabel()` method above will fallback to returning the URI
+                // if no label was found. We don't want that here.
+                $ret['to']['memberSet'][0]['prefLabel'] = [
+                    $labelLang => $labelValue,
+                ];
+            }
+        }
+
+        return $ret;
+    }
+
 }
+
