@@ -1,162 +1,280 @@
 <?php
 
 /**
- * GlobalConfig provides access to the Skosmos configuration in config.inc.
+ * Setting some often needed namespace prefixes
  */
-class GlobalConfig {
-    private $languages;
+EasyRdf\RdfNamespace::set('skosmos', 'http://purl.org/net/skosmos#');
+EasyRdf\RdfNamespace::set('skosext', 'http://purl.org/finnonto/schema/skosext#');
+EasyRdf\RdfNamespace::set('isothes', 'http://purl.org/iso25964/skos-thes#');
+EasyRdf\RdfNamespace::set('mads', 'http://www.loc.gov/mads/rdf/v1#');
+EasyRdf\RdfNamespace::set('wd', 'http://www.wikidata.org/entity/');
+EasyRdf\RdfNamespace::set('wdt', 'http://www.wikidata.org/prop/direct/');
 
-    public function __construct($config_name='/../config.inc') 
+/**
+ * GlobalConfig provides access to the Skosmos configuration in config.ttl.
+ */
+class GlobalConfig extends BaseConfig {
+
+    /** Cache reference */
+    private $cache;
+    /** Location of the configuration file. Used for caching. */
+    private $filePath;
+    /** Namespaces from vocabularies configuration file. */
+    private $namespaces;
+    /** EasyRdf\Graph graph */
+    private $graph;
+    /**
+     * @var int the time the config file was last modified
+     */
+    private $configModifiedTime = null;
+
+    public function __construct($config_name='/../config.ttl')
     {
+        $this->cache = new Cache();
         try {
-            $file_path = dirname(__FILE__) . $config_name;
-            if (!file_exists($file_path)) {
-                throw new Exception('config.inc file is missing, please provide one.');
+            $this->filePath = realpath( dirname(__FILE__) . $config_name );
+            if (!file_exists($this->filePath)) {
+                throw new Exception('config.ttl file is missing, please provide one.');
             }
-            require_once($file_path);
-            if (isset($LANGUAGES)) {
-                $this->languages = $LANGUAGES;
-            }
+            $this->initializeConfig();
         } catch (Exception $e) {
             echo "Error: " . $e->getMessage();
             return;
         }
     }
-    
-    private function getConstant($name, $default)
+
+    public function getCache()
     {
-        if (defined($name) && constant($name)) {
-            return constant($name);
+        return $this->cache;
+    }
+
+    /**
+     * @return int the time the config file was last modified
+     */
+    public function getConfigModifiedTime()
+    {
+        return $this->configModifiedTime;
+    }
+
+    /**
+     * Initialize configuration, reading the configuration file from the disk,
+     * and creating the graph and resources objects. Uses a cache if available,
+     * in order to avoid re-loading the complete configuration on each request.
+     */
+    private function initializeConfig()
+    {
+        try {
+            // retrieve last modified time for config file (filemtime returns int|bool!)
+            $configModifiedTime = filemtime($this->filePath);
+            if (!is_bool($configModifiedTime)) {
+                $this->configModifiedTime = $configModifiedTime;
+            }
+            // use APC user cache to store parsed config.ttl configuration
+            if ($this->cache->isAvailable() && !is_null($this->configModifiedTime)) {
+                // @codeCoverageIgnoreStart
+                $key = realpath($this->filePath) . ", " . $this->configModifiedTime;
+                $nskey = "namespaces of " . $key;
+                $this->graph = $this->cache->fetch($key);
+                $this->namespaces = $this->cache->fetch($nskey);
+                if ($this->graph === false || $this->namespaces === false) { // was not found in cache
+                    $this->parseConfig($this->filePath);
+                    $this->cache->store($key, $this->graph);
+                    $this->cache->store($nskey, $this->namespaces);
+                }
+                // @codeCoverageIgnoreEnd
+            } else { // APC not available, parse on every request
+                $this->parseConfig($this->filePath);
+            }
+
+            $configResources = $this->graph->allOfType("skosmos:Configuration");
+            if (is_null($configResources) || !is_array($configResources) || count($configResources) !== 1) {
+                throw new Exception("config.ttl must have exactly one skosmos:Configuration");
+            }
+
+            $this->resource = $configResources[0];
+            $this->initializeNamespaces();
+        } catch (Exception $e) {
+            echo "Error: " . $e->getMessage();
+        }      
+    }
+
+    /**
+     * Parses configuration from the config.ttl file
+     * @param string $filename path to config.ttl file
+     * @throws \EasyRdf\Exception
+     */
+    private function parseConfig($filename)
+    {
+        $this->graph = new EasyRdf\Graph();
+        $parser = new SkosmosTurtleParser();
+        $parser->parse($this->graph, file_get_contents($filename), 'turtle', $filename);
+        $this->namespaces = $parser->getNamespaces();
+    }
+
+    /**
+     * Returns the graph created after parsing the configuration file.
+     * @return \EasyRdf\Graph
+     */
+    public function getGraph()
+    {
+        return $this->graph;
+    }
+
+    /**
+     * Registers RDF namespaces from the config.ttl file for use by EasyRdf (e.g. serializing)
+     */
+    private function initializeNamespaces() {
+        foreach ($this->namespaces as $prefix => $fullUri) {
+            if ($prefix != '' && EasyRdf\RdfNamespace::get($prefix) === null) // if not already defined
+            {
+                EasyRdf\RdfNamespace::set($prefix, $fullUri);
+            }
         }
-        return $default;
     }
 
     /**
      * Returns the UI languages specified in the configuration or defaults to
      * only show English
-     * @return array 
+     * @return array
      */
-    public function getLanguages() 
+    public function getLanguages()
     {
-        if ($this->languages) {
-            return $this->languages;
+        $languageResources = $this->getResource()->getResource('skosmos:languages');
+        if (!is_null($languageResources) && !empty($languageResources)) {
+            $languages = array();
+            foreach ($languageResources as $languageResource) {
+                /** @var \EasyRdf\Literal $languageName */
+                $languageName = $languageResource->getLiteral('rdfs:label');
+                /** @var \EasyRdf\Literal $languageValue */
+                $languageValue = $languageResource->getLiteral('rdf:value');
+                if ($languageName && $languageValue) {
+                    $languages[$languageName->getValue()] = $languageValue->getValue();
+                }
+            }
+            return $languages;
+        } else {
+            return array('en' => 'en_GB.utf8');
         }
-        return array('en' => 'en_GB.utf8');
     }
-    
-    /**
-     * Returns the vocabulary configuration file specified the configuration
-     * or vocabularies.ttl if not found.
-     * @return string
-     */
-    public function getVocabularyConfigFile() 
-    {
-        return $this->getConstant('VOCABULARIES_FILE', 'vocabularies.ttl');
-    }
-    
+
     /**
      * Returns the external HTTP request timeout in seconds or the default value
      * of 5 seconds if not specified in the configuration.
      * @return integer
      */
-    public function getHttpTimeout() 
+    public function getHttpTimeout()
     {
-        return $this->getConstant('HTTP_TIMEOUT', 5);
+        return $this->getLiteral('skosmos:httpTimeout', 5);
     }
-    
+
     /**
      * Returns the SPARQL HTTP request timeout in seconds or the default value
      * of 20 seconds if not specified in the configuration.
      * @return integer
      */
-    public function getSparqlTimeout() 
+    public function getSparqlTimeout()
     {
-        return $this->getConstant('SPARQL_TIMEOUT', 20);
+        return $this->getLiteral('skosmos:sparqlTimeout', 20);
     }
-    
+
     /**
      * Returns the sparql endpoint address defined in the configuration. If
-     * not then defaulting to http://localhost:3030/ds/sparql 
+     * not then defaulting to http://localhost:3030/ds/sparql
      * @return string
      */
-    public function getDefaultEndpoint() 
+    public function getDefaultEndpoint()
     {
-        return $this->getConstant('DEFAULT_ENDPOINT', 'http://localhost:3030/ds/sparql');
+        $endpoint = $this->resource->get('skosmos:sparqlEndpoint');
+        if ($endpoint) {
+            return $endpoint->getUri();
+        } else {
+            return 'http://localhost:3030/ds/sparql';
+        }
     }
-    
-    /**
-     * @return string
-     */
-    public function getSparqlGraphStore() 
-    {
-        return $this->getConstant('SPARQL_GRAPH_STORE', null);
-    }
-    
+
     /**
      * Returns the maximum number of items to return in transitive queries if defined
      * in the configuration or the default value of 1000.
-     * @return integer 
+     * @return integer
      */
-    public function getDefaultTransitiveLimit() 
+    public function getDefaultTransitiveLimit()
     {
-        return $this->getConstant('DEFAULT_TRANSITIVE_LIMIT', 1000);
+        return $this->getLiteral('skosmos:transitiveLimit', 1000);
     }
-    
+
     /**
      * Returns the maximum number of items to load at a time if defined
      * in the configuration or the default value of 20.
-     * @return integer 
+     * @return integer
      */
-    public function getSearchResultsSize() 
+    public function getSearchResultsSize()
     {
-        return $this->getConstant('SEARCH_RESULTS_SIZE', 20);
+        return $this->getLiteral('skosmos:searchResultsSize', 20);
     }
-    
+
     /**
      * Returns the configured location for the twig template cache and if not
      * defined defaults to "/tmp/skosmos-template-cache"
      * @return string
      */
-    public function getTemplateCache() 
+    public function getTemplateCache()
     {
-        return $this->getConstant('TEMPLATE_CACHE', '/tmp/skosmos-template-cache');
+        return $this->getLiteral('skosmos:templateCache', '/tmp/skosmos-template-cache');
     }
-    
+
     /**
-     * Returns the defined sparql-query extension eg. "JenaText" or 
+     * Returns the defined sparql-query extension eg. "JenaText" or
      * if not defined falling back to SPARQL 1.1
      * @return string
      */
-    public function getDefaultSparqlDialect() 
+    public function getDefaultSparqlDialect()
     {
-        return $this->getConstant('DEFAULT_SPARQL_DIALECT', 'Generic');
+        return $this->getLiteral('skosmos:sparqlDialect', 'Generic');
     }
 
     /**
      * Returns the feedback address defined in the configuration.
      * @return string
      */
-    public function getFeedbackAddress() 
+    public function getFeedbackAddress()
     {
-        return $this->getConstant('FEEDBACK_ADDRESS', null);
+        return $this->getLiteral('skosmos:feedbackAddress', null);
     }
-    
+
+    /**
+     * Returns the feedback sender address defined in the configuration.
+     * @return string
+     */
+    public function getFeedbackSender()
+    {
+        return $this->getLiteral('skosmos:feedbackSender', null);
+    }
+
+    /**
+     * Returns the feedback envelope sender address defined in the configuration.
+     * @return string
+     */
+    public function getFeedbackEnvelopeSender()
+    {
+        return $this->getLiteral('skosmos:feedbackEnvelopeSender', null);
+    }
+
     /**
      * Returns true if exception logging has been configured.
-     * @return boolean 
+     * @return boolean
      */
-    public function getLogCaughtExceptions() 
+    public function getLogCaughtExceptions()
     {
-        return $this->getConstant('LOG_CAUGHT_EXCEPTIONS', FALSE);
+        return $this->getBoolean('skosmos:logCaughtExceptions', FALSE);
     }
-    
+
     /**
      * Returns true if browser console logging has been enabled,
      * @return boolean
      */
     public function getLoggingBrowserConsole()
     {
-        return $this->getConstant('LOG_BROWSER_CONSOLE', FALSE);
+        return $this->getBoolean('skosmos:logBrowserConsole', FALSE);
     }
 
     /**
@@ -165,71 +283,62 @@ class GlobalConfig {
      */
     public function getLoggingFilename()
     {
-        return $this->getConstant('LOG_FILE_NAME', null);
+        return $this->getLiteral('skosmos:logFileName', null);
     }
 
     /**
      * @return string
      */
-    public function getServiceName() 
+    public function getServiceName()
     {
-        return $this->getConstant('SERVICE_NAME', 'Skosmos');
+        return $this->getLiteral('skosmos:serviceName', 'Skosmos');
     }
-    
+
     /**
      * @return string
      */
-    public function getServiceTagline() 
+    public function getCustomCss()
     {
-        return $this->getConstant('SERVICE_TAGLINE', null);
+        return $this->getLiteral('skosmos:customCss', null);
     }
-    
-    /**
-     * @return string
-     */
-    public function getServiceLogo() 
-    {
-        return $this->getConstant('SERVICE_LOGO', null);
-    }
-    
-    /**
-     * @return string
-     */
-    public function getCustomCss() 
-    {
-        return $this->getConstant('CUSTOM_CSS', null);
-    }
-    
+
     /**
      * @return boolean
      */
-    public function getUiLanguageDropdown() 
+    public function getUiLanguageDropdown()
     {
-        return $this->getConstant('UI_LANGUAGE_DROPDOWN', FALSE);
+        return $this->getBoolean('skosmos:uiLanguageDropdown', FALSE);
     }
-    
+
     /**
      * @return string
      */
-    public function getBaseHref() 
+    public function getBaseHref()
     {
-        return $this->getConstant('BASE_HREF', null);
+        return $this->getLiteral('skosmos:baseHref', null);
     }
-    
+
     /**
-     * @return string
+     * @return array
      */
-    public function getGlobalPlugins() 
+    public function getGlobalPlugins()
     {
-        return explode(' ', $this->getConstant('GLOBAL_PLUGINS', null));
+        $globalPlugins = array();
+        $globalPluginsResource =  $this->getResource()->getResource("skosmos:globalPlugins");
+        if ($globalPluginsResource) {
+            foreach ($globalPluginsResource as $resource) {
+                $globalPlugins[] = $resource->getValue();
+            }
+        }
+        return $globalPlugins;
     }
-    
+
     /**
      * @return boolean
      */
     public function getHoneypotEnabled()
     {
-        return $this->getConstant('UI_HONEYPOT_ENABLED', TRUE);
+        return $this->getBoolean('skosmos:uiHoneypotEnabled', TRUE);
     }
 
     /**
@@ -237,7 +346,7 @@ class GlobalConfig {
      */
     public function getHoneypotTime()
     {
-        return $this->getConstant('UI_HONEYPOT_TIME', 5);
+        return $this->getLiteral('skosmos:uiHoneypotTime', 5);
     }
 
     /**
@@ -245,6 +354,6 @@ class GlobalConfig {
      */
     public function getCollationEnabled()
     {
-        return $this->getConstant('SPARQL_COLLATION_ENABLED', FALSE);
+        return $this->getBoolean('skosmos:sparqlCollationEnabled', FALSE);
     }
 }
