@@ -1,13 +1,14 @@
 <?php
 
 /**
- * VocabularyConfig provides access to the vocabulary configuration defined in vocabularies.ttl.
+ * VocabularyConfig provides access to the vocabulary configuration defined in config.ttl.
  */
-class VocabularyConfig extends DataObject
+class VocabularyConfig extends BaseConfig
 {
     private $plugins;
+    private $languageOrderCache = array();
 
-    public function __construct($resource, $globalPlugins=array()) 
+    public function __construct($resource, $globalPlugins=array())
     {
         $this->resource = $resource;
         $plugins = $this->resource->allLiterals('skosmos:usePlugin');
@@ -18,43 +19,6 @@ class VocabularyConfig extends DataObject
             }
         }
         $this->plugins = new PluginRegister(array_merge($globalPlugins, $pluginArray));
-    }
-
-    /**
-     * Returns a boolean value based on a literal value from the vocabularies.ttl configuration.
-     * @param string $property the property to query
-     * @param boolean $default the default value if the value is not set in configuration
-     */
-    private function getBoolean($property, $default = false)
-    {
-        $val = $this->resource->getLiteral($property);
-        if ($val) {
-            return filter_var($val->getValue(), FILTER_VALIDATE_BOOLEAN);
-        }
-
-        return $default;
-    }
-    
-    /**
-     * Returns a boolean value based on a literal value from the vocabularies.ttl configuration.
-     * @param string $property the property to query
-     * @param string $lang preferred language for the literal,
-     */
-    private function getLiteral($property, $lang=null)
-    {
-        if (!isset($lang)) {;
-            $lang = $this->getEnvLang();
-        }
-
-        $literal = $this->resource->getLiteral($property, $lang);
-        if ($literal) {
-            return $literal->getValue();
-        }
-
-        // not found with selected language, try any language
-        $literal = $this->resource->getLiteral($property);
-        if ($literal)
-          return $literal->getValue();
     }
 
     /**
@@ -95,7 +59,7 @@ class VocabularyConfig extends DataObject
     public function getShortName()
     {
         $shortname = $this->getLiteral('skosmos:shortName');
-        if ($shortname) 
+        if ($shortname)
           return $shortname;
 
         // if no shortname exists fall back to the id
@@ -119,20 +83,20 @@ class VocabularyConfig extends DataObject
      */
     public function getTitle($lang = null)
     {
-        return $this->getLiteral('dc:title', $lang);
+        return $this->getLiteral('dc:title', false, $lang);
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function sortByNotation()
     {
         return $this->getBoolean('skosmos:sortByNotation');
     }
-    
+
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function showChangeList()
@@ -170,6 +134,22 @@ class VocabularyConfig extends DataObject
             $ret[$mimetype] = $url->getURI();
         }
         return $ret;
+    }
+
+    /**
+     * Returns the main Concept Scheme URI of that Vocabulary,
+     * or null if not set.
+     * @return string concept scheme URI or null
+     */
+
+    public function getMainConceptSchemeURI()
+    {
+        $val = $this->resource->getResource("skosmos:mainConceptScheme");
+        if ($val) {
+            return $val->getURI();
+        }
+
+        return null;
     }
 
     /**
@@ -249,7 +229,7 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function getShowHierarchy()
@@ -258,7 +238,7 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function showConceptSchemesInHierarchy()
@@ -267,7 +247,7 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean defaults to true if fetching hasn't been explicitly denied.
      */
     public function getExternalResourcesLoading()
@@ -276,7 +256,7 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function getShowLangCodes()
@@ -285,17 +265,30 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns skosmos:marcSourcecode value set in config.ttl.
+     * @return string marcsource name
+     */
+    public function getMarcSourceCode($lang = null)
+    {
+        return $this->getLiteral('skosmos:marcSourceCode', false, $lang);
+    }
+
+    /**
+     * Returns a boolean value set in the config.ttl config.
      * @return array array of concept class URIs (can be empty)
      */
     public function getIndexClasses()
     {
-        $resources = $this->resource->allResources("skosmos:indexShowClass");
-        $ret = array();
-        foreach ($resources as $res) {
-            $ret[] = $res->getURI();
-        }
-        return $ret;
+        return $this->getResources("skosmos:indexShowClass");
+    }
+
+    /**
+     * Returns skosmos:externalProperty values set in the config.ttl config.
+     * @return array array of external property URIs (can be empty)
+     */
+    public function getExtProperties()
+    {
+        return $this->getResources("skosmos:externalProperty");
     }
 
     /**
@@ -329,7 +322,7 @@ class VocabularyConfig extends DataObject
             }
 
         }
-        if ($this->showAlphabeticalIndex() === false) { 
+        if ($this->showAlphabeticalIndex() === false) {
             if ($this->getShowHierarchy()) {
                 return 'hierarchy';
             } else if ($this->getGroupClassURI()) {
@@ -378,7 +371,7 @@ class VocabularyConfig extends DataObject
         $ret = array();
         foreach ($resources as $res) {
             $prop = $res->getURI();
-            if (EasyRdf\RdfNamespace::shorten($prop) !== null) // prefixing if possible 
+            if (EasyRdf\RdfNamespace::shorten($prop) !== null) // prefixing if possible
             {
                 $prop = EasyRdf\RdfNamespace::shorten($prop);
             }
@@ -389,7 +382,7 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function showNotation()
@@ -398,16 +391,16 @@ class VocabularyConfig extends DataObject
     }
 
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function showAlphabeticalIndex()
     {
         return $this->getBoolean('skosmos:showAlphabeticalIndex', true);
     }
-    
+
     /**
-     * Returns a boolean value set in the vocabularies.ttl config.
+     * Returns a boolean value set in the config.ttl config.
      * @return boolean
      */
     public function getShowDeprecated()
@@ -429,5 +422,47 @@ class VocabularyConfig extends DataObject
             $ret[] = array('uri' => $prop, 'prefLabel' =>  $label->getValue());
         }
         return $ret;
+    }
+
+    /**
+     * Returns an array of fallback languages that is ordered by priority and
+     * defined in the vocabulary configuration as a collection.
+     * Additionally, the chosen content language is inserted with the highest priority
+     * and the vocab default language is inserted with the lowest priority.
+     * @param string $clang
+     * @return array of language code strings
+     */
+    public function getLanguageOrder($clang)
+    {
+        if (array_key_exists($clang, $this->languageOrderCache)) {
+            return $this->languageOrderCache[$clang];
+        }
+        $ret = array($clang);
+        $fallbacks = !empty($this->resource->get('skosmos:fallbackLanguages')) ? $this->resource->get('skosmos:fallbackLanguages') : array();
+        foreach ($fallbacks as $lang) {
+            if (!in_array($lang, $ret)) {
+                $ret[] = (string)$lang; // Literal to string conversion
+            }
+        }
+        if (!in_array($this->getDefaultLanguage(), $ret)) {
+            $ret[] = (string)$this->getDefaultLanguage();
+        }
+        foreach ($this->getLanguages() as $lang) {
+            if (!in_array($lang, $ret)) {
+                $ret[] = $lang;
+            }
+        }
+        // store in cache so this doesn't have to be computed again
+        $this->languageOrderCache[$clang] = $ret;
+        return $ret;
+    }
+
+    /**
+     * Returns a boolean value set in the config.ttl config.
+     * @return boolean
+     */
+    public function getUseModifiedDate(): bool
+    {
+        return $this->getBoolean('skosmos:useModifiedDate', false);
     }
 }
