@@ -1,3 +1,3 @@
 <?php
-require_once('vendor/autoload.php');
-require_once('model/Model.php');
+require_once(__DIR__ . '/../vendor/autoload.php');
+require_once(__DIR__ . '/../model/Model.php');
